Remove childNodes that bertram didnt use
Lines 1306-1311

Added a loop that removes the childnodes of an element.
Bertram did not organize his elements into an hierarchy and didn't use
many divs or section-tags. I did, so I have to add them to the removal
process of elements, that are not needed anymore during the rebuilding
of the page